feat(core): widen length path to 32-bit, lifting the 64KB SDU cap
The reassembly buffer, transport, and request/response dispatch were
uint16_t end-to-end, capping a single ISO-TP SDU at 65535 bytes. With
the escape FirstFrame now carrying a 32-bit FF_DL, widen the whole
length path to uint32_t so large CAN-FD transfers work end to end:

- transport: msg_len, bytes_processed, tx_sdu_size/len, send length
- core: uds_input_sdu, handle_request/execute_handler, send_response,
  client_request, parse_addr_len, and the truncating client-cb cast
- config: rx_buffer_size, tx_buffer_size, and the length-bearing
  callbacks (tp_send, service handler, did_write, response_cb, is_safe)

BREAKING: integrators implementing the widened callbacks must change
their length parameter from uint16_t to uint32_t. DID/security buffer
capacities and resp_len are unchanged. In-repo examples and tests
updated to match.

Tested: new >64KB escape-FF TX/RX reassembly cases (70000-byte SDU)
plus the existing escape/FC suite; full suite 45/45 green.

+++ examples/zephyr_uds_server/src/main.c
@@ -14,11 +14,11 @@ extern void uds_log_zephyr(uint8_t level, const char* msg);
 
 #if defined(CONFIG_UDSLIB_TRANSPORT_NATIVE)
 extern int uds_zephyr_isotp_init(uint32_t rx_id, uint32_t tx_id);
-extern int uds_zephyr_isotp_send(struct uds_ctx* ctx, const uint8_t* data, uint16_t len);
+extern int uds_zephyr_isotp_send(struct uds_ctx* ctx, const uint8_t* data, uint32_t len);
 extern int uds_zephyr_isotp_recv(uint8_t* buf, uint16_t size);
 #elif defined(CONFIG_UDSLIB_TRANSPORT_FALLBACK)
 extern int uds_zephyr_tp_fallback_init(struct uds_ctx* uds_ctx, uint32_t rx_id, uint32_t tx_id);
-extern int uds_zephyr_tp_fallback_send(struct uds_ctx* uds_ctx, const uint8_t* data, uint16_t len);
+extern int uds_zephyr_tp_fallback_send(struct uds_ctx* uds_ctx, const uint8_t* data, uint32_t len);
 extern void uds_zephyr_tp_fallback_process(uint32_t time_ms);
 #endif
